refactor to reuse components
